+++ docs/CHANGELOG.rst
@@ -16,6 +16,8 @@ Added
 
 Changed
 -------
+- **BACKWARD INCOMPATIBLE:** Implement peak calling step and
+  add QC reporting to the Cut & Run workflow
 - Bump requirements versions
 
 Fixed